requirements.txt
# LangChain core
langchain>=1.2.13
langchain-core>=1.2.22
langchain-openai>=1.1.12
langchain-community>=0.4.1
langchain-text-splitters>=1.1.1

# LangChain integrations
langchain-classic>=1.0.3
langchain-chroma>=1.1.0

# Vector store
chromadb>=1.5.5

# Document loaders
pypdf>=6.9.2
beautifulsoup4>=4.14.3
lxml>=6.0.2

# OpenAI
openai>=2.30.0

# Data validation
pydantic>=2.12.5

# Environment variables
python-dotenv>=1.2.2
